Renamed TxResult to QueryResult and moved it from tx.rs to rpc.rs where other queries are. Improved formatting when printing transaction results.
